requirements.txt
asgiref==3.5.2
cement==3.0.8
certifi==2022.9.14
charset-normalizer==2.1.1
colorama==0.4.5
commonmark==0.9.1
defusedxml==0.7.1
Django==4.1.1
django-municipios==1.2.1
djangorestframework==3.14.0
future==0.18.2
httpie==3.2.1
jmespath==1.0.1
multidict==6.0.2
mysqlclient==2.1.1
pathspec==0.10.1
Pygments==2.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
semantic-version==2.10.0
six==1.16.0
sqlparse==0.4.2
termcolor==2.0.1
tzdata==2022.2
urllib3==1.26.12